Newer architecture with a lot of fixes
