+++ middleware/api_handlers.go
@@ -1,0 +1,13 @@
+package middleware
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+)
+
+func MainAPIHandler(c *gin.Context) {
+	c.JSON(http.StatusOK, gin.H{
+		"message": "Pong!",
+	})
+}
