+++ repository/database.go
@@ -1,4 +1,4 @@
-package database
+package repository
 
 import (
 	"database/sql"
